pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.3"
pytest = "^7.2.1"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
validators = "^0.20.0"
psycopg2 = "^2.9.5"
psycopg2-binary = "^2.9.5"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
